requirements.txt
python-socketio[client] ~= 5.0.0
pylint ~= 2.6.0
autopep8 ~= 1.5.4
numpy == 1.19.3
scikit-learn ~= 0.24.0
pandas ~= 1.2.0
joblib ~= 1.0.0
requests ~= 2.25.1
